Move things to a papertoaster module
