--- artworks/__init__.py
@@ -1,29 +1,0 @@
-from artworks.barcode_code128 import Barcode128
-from artworks.braids import Braids
-from artworks.colored_braids import ColoredBraids
-from artworks.edge_directions import EdgeDirectionTiling
-from artworks.elementary_ca import ElementaryCA
-from artworks.grid import Grid
-from artworks.hex_grid import HexGrid
-from artworks.hitomezashi import Hitomezashi
-from artworks.iso_grid import IsoGrid
-from artworks.quiet_dice import QuietDice
-from artworks.robot_walks import RobotWalks
-from artworks.todo import ToDoList
-from artworks.turtle_dance import TurtleDance
-
-ARTWORKS = [
-    Barcode128,
-    Braids,
-    ColoredBraids,
-    EdgeDirectionTiling,
-    ElementaryCA,
-    Grid,
-    HexGrid,
-    Hitomezashi,
-    IsoGrid,
-    QuietDice,
-    RobotWalks,
-    ToDoList,
-    TurtleDance
-]
