+++ src/papertoaster/artworks/__init__.py
@@ -1,0 +1,29 @@
+from papertoaster.artworks.barcode_code128 import Barcode128
+from papertoaster.artworks.braids import Braids
+from papertoaster.artworks.colored_braids import ColoredBraids
+from papertoaster.artworks.edge_directions import EdgeDirectionTiling
+from papertoaster.artworks.elementary_ca import ElementaryCA
+from papertoaster.artworks.grid import Grid
+from papertoaster.artworks.hex_grid import HexGrid
+from papertoaster.artworks.hitomezashi import Hitomezashi
+from papertoaster.artworks.iso_grid import IsoGrid
+from papertoaster.artworks.quiet_dice import QuietDice
+from papertoaster.artworks.robot_walks import RobotWalks
+from papertoaster.artworks.todo import ToDoList
+from papertoaster.artworks.turtle_dance import TurtleDance
+
+ARTWORKS = [
+    Barcode128,
+    Braids,
+    ColoredBraids,
+    EdgeDirectionTiling,
+    ElementaryCA,
+    Grid,
+    HexGrid,
+    Hitomezashi,
+    IsoGrid,
+    QuietDice,
+    RobotWalks,
+    ToDoList,
+    TurtleDance
+]
