+++ src/papertoaster/artworks/barcode_code128.py
@@ -1,11 +1,10 @@
-import argparse
-
-from postscriptlib import receipts
-from postscriptlib.barcodes.code128 import Code128
-from postscriptlib.path import Path
+from papertoaster import receipts
+from papertoaster.barcodes.code128 import Code128
+from papertoaster.path import Path
 
 # Postscript points-per-inch
 PPI = receipts.Receipt.PPI
+
 
 class Barcode128(receipts.Receipt):
     ARTWORK_ID = 'barcode128'
@@ -23,10 +22,11 @@ class Barcode128(receipts.Receipt):
         module_width = 1
         module_height = PPI
         self.barcode = Code128(module_width, module_height)
-    
+
     def draw(self):
         defines = self.barcode.make_definitions()
-        barcode_lines = self.barcode.draw(0, 0, self.width, self.height, self.text)
+        barcode_lines = self.barcode.draw(
+            0, 0, self.width, self.height, self.text)
 
         self.postscript_lines.extend(defines + barcode_lines)
