+++ src/papertoaster/barcodes/barcode.py
